requirements.txt
matplotlib>=3.4.3
numpy>=1.19.5
scikit-image>=0.19.1
PyYAML>=6.0
easydict>=1.9
pandas>=1.3.3
sklearn>=0.0
scikit-learn>=0.24.2
Pillow>=8.3.2
opencv-python>=4.5.5.62
scipy>=1.7.1
einops>=0.4.0
tensorboardX>=2.5